# functions/openapi_server/openapi/openapi.yaml
---
openapi: 3.0.1
info:
  title: restingest
  description: restingest generic endpoints
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /
paths:
  /generic:
    post:
      summary: Store some data
      description: Store the posted data
      operationId: generic_post
      requestBody:
        description: A new JSON document to be stored.
        content:
          application/json: {}
        required: true
      responses:
        201:
          description: Successful response.
      x-codegen-request-body-name: body
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /urencnt:
    post:
      summary: Store some data
      description: Store the posted data
      operationId: urencnt_post
      requestBody:
        description: A new JSON document to be stored.
        content:
          application/json: {}
        required: true
      responses:
        201:
          description: Successful response.
      x-codegen-request-body-name: body
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /urenover:
    post:
      summary: Store some data
      description: Store the posted data
      operationId: urenover_post
      requestBody:
        description: A new JSON document to be stored.
        content:
          application/json: {}
        required: true
      responses:
        201:
          description: Successful response.
      x-codegen-request-body-name: body
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /urendesc:
    post:
      summary: Store some data
      description: Store the posted data
      operationId: urendesc_post
      requestBody:
        description: A new JSON document to be stored.
        content:
          application/json: {}
        required: true
      responses:
        201:
          description: Successful response.
      x-codegen-request-body-name: body
      x-openapi-router-controller: openapi_server.controllers.default_controller